FLINK-10114: Move the implementation to flink-orc
